requirements.txt
# Core
requests>=2.31.0
urllib3>=2.0

# PDF Analysis
PyMuPDF>=1.23.0
pdfminer.six>=20221105

# NLP
nltk>=3.8
scikit-learn>=1.3.0

# Graph
networkx>=3.1

# Visualization
matplotlib>=3.8.0
plotly>=5.18.0

# Interactive Dashboard
dash>=2.14.0

# Database (stdlib sqlite3 – no extra install needed)
# IPFS (uses Kubo HTTP RPC API via requests – no extra packages needed)

# Utilities
python-dateutil>=2.8.0
